test(engine): expand enrichment and deep integration tests across 26 modules
Add and update integration tests across the test suite to improve
coverage depth for recently hardened modules:

- ambient_authority_integration: File::/OpenOptions:: pattern tests
- anti_entropy_enrichment: expanded trust state coverage
- array_fast_lane_integration: double-deopt idempotency, reopt cycle
- bayesian_error_recovery_enrichment: additional recovery paths
- benchmark_coverage_saturation_enrichment: saturation thresholds
- bifurcation_boundary_scanner_enrichment: boundary detection cases
- calibration_sentinel_enrichment: sentinel trigger sequences
- claim_atom_lattice_enrichment: lattice join/meet operations
- counterfactual_evaluator_enrichment: evaluation edge cases
- declassification_pipeline_integration: receipt signing order,
  stats_at boundary, emergency grant expiry validation
- descent_certificate_gate_enrichment: gate verdict paths
- kernel_shift_monitor_deep: monitor trigger sequences
- metadata_substrate_governance_enrichment: governance rules
- offline_synthesis_pipeline_deep: pipeline stage transitions
- persistent_cache_contract_enrichment: cache invalidation
- phase_gate_enrichment: phase transition coverage
- promise_model_enrichment: promise resolution paths
- queueing_admission_control (deep + enrichment): admission rules
- remote_computation_registry_enrichment: registry operations
- resource_certificate_governance (deep + enrichment): cert lifecycle
- session_hostcall_channel_deep: channel protocol tests
- shipped_path_parity_cli: mismatch catalog validation
- version_matrix_lane_enrichment: version resolution
- workload_corpus_gate_integration: minor consistency fix

+++ crates/franken-engine/tests/promise_model_enrichment_integration.rs
@@ -769,7 +769,7 @@ fn microtask_queue_deterministic_ordering() {
             q.enqueue(Microtask::PromiseReaction {
                 handler: Some(ClosureHandle(i)),
                 argument: js_int(i as i64),
-                result_promise: PromiseHandle(i as u32),
+                result_promise: PromiseHandle(i),
                 label: Label::Public,
             });
         }
